Reorganization Update
-renamed testing directory to test
-updated build process so that KMC_Lattice is separately built using its
own makefile into an archive and is linked into Excimontec during its
build
-KMC_Lattice is now tested separately and not inlcuded in the tests and
coverage of Excimontec
-updated makefile to use different flags depending on the build target,
with options for test_coverage and test targets
-updated travis yml file to specify different build targets so that the
slower coverage test is done separately from the compiler build tests
